\name{BayesianMediationA-package}
\alias{BayesianMediationA-package}
\alias{BayesianMediationA}
\docType{package}
\title{
Bayesian Mediation Analysis Package
}
\description{
Functions in the package are used for Bayesian mediation analysis. Mediation effect refers to the effect conveyed by intervening variables to an observed relationship between an exposure and a response variable (outcome).  In this package, the exposure is called the predictor, the intervening variables are called mediators.  The mediation effects include the total effect, direct effect, and indirect effect, which are defined and the statistical inferences described in Yu et al.(2014). We use four different method for the mediation analysis (Yu et al. 2022).
}
\author{
Qingzhao Yu  \email{qyu@lsuhsc.edu} and Bin Li \email{bli@lsu.edu}

Maintainer: Qingzhao Yu  \email{qyu@lsuhsc.edu}
}
\references{
Yu, Q., Fan, Y., and Wu, X. (2014) <doi:10.4172/2155-6180.1000189>. General Multiple Mediation Analysis With an Application to Explore Racial Disparity in Breast Cancer Survival, Journal of Biometrics and Biostatistics, 5(2):189.

Yu, Q. and Li, B., 2022. Statistical Methods for Mediation, Confounding and Moderation Analysis Using R and SAS. Chapman and Hall/CRC. ISBN 9780367365479.
}
