\name{print.summary.bma}
\alias{print.summary.bma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print the results from the summary function.
}
\description{
Print the summary results from summary.bma.bx.cy.
}
\usage{
\method{print}{summary.bma}(x, ..., digit = x$digit, method = x$method, RE = x$RE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
the oject from summary.bma.bx.cy.
}
  \item{\dots}{
other arguments.
}
  \item{digit}{
the nubmer of digits to be printed.
}
  \item{method}{
the method to be used.
}
  \item{RE}{
if ture, print the relative effect
}
}
\value{
Print and plot from the summary of bma.bx.cy.
}
\author{
Qingzhao Yu and Bin Li
}
\examples{
# see example at bma.bx.cy.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
